pyproject.toml
[project]
name = "tg-w-t-mar29-2026"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "paho-mqtt>=2.1.0",
    "python-telegram-bot>=22.7",
    "rpi-gpio>=0.7.1",
    "spidev>=3.8",
]
